+++ packages/mime/scripts/codegen.d.ts
@@ -1,0 +1,3 @@
+export function generateCompressibleMimeTypesContent(): string
+export function generateMimeTypesContent(): string
+