+++ packages/mime/src/index.ts
@@ -1,0 +1,3 @@
+export { detectMimeType } from './lib/detect-mime-type.ts'
+export { isCompressibleMimeType } from './lib/is-compressible-mime-type.ts'
+